Update tests to pass
I could use some genuine unit tests on the getGasPrice method
and more thorough integration testing of the behavior where
the input field NOT modified most recently is the one the
app updates
